back up copy of new post modal and integrate image picker with modal
